Print correctly expanded variable with "". Also add file for checking first argv in input
